pyproject.toml
[project]
name = "generic-paint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "imageio",
    "jupyter",
    "matplotlib",
    "nbconvert",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "plotly",
    "seaborn",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
    "transformers>=4.56.0",
    "uvicorn>=0.35.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu124"
